+++ mobile/android/android-components/components/support/base/src/main/java/mozilla/components/support/base/log/logger/Logger.kt
@@ -18,28 +18,28 @@ class Logger(
     /**
      * Send a DEBUG log message.
      */
-    fun debug(message: String? = null, throwable: Throwable? = null) {
+    fun debug(message: String, throwable: Throwable? = null) {
         Log.log(Log.Priority.DEBUG, tag = tag, message = message, throwable = throwable)
     }
 
     /**
      * Send a INFO log message.
      */
-    fun info(message: String? = null, throwable: Throwable? = null) {
+    fun info(message: String, throwable: Throwable? = null) {
         Log.log(Log.Priority.INFO, tag = tag, message = message, throwable = throwable)
     }
 
     /**
      * Send a WARN log message.
      */
-    fun warn(message: String? = null, throwable: Throwable? = null) {
+    fun warn(message: String, throwable: Throwable? = null) {
         Log.log(Log.Priority.WARN, tag = tag, message = message, throwable = throwable)
     }
 
     /**
      * Send a ERROR log message.
      */
-    fun error(message: String? = null, throwable: Throwable? = null) {
+    fun error(message: String, throwable: Throwable? = null) {
         Log.log(Log.Priority.ERROR, tag = tag, message = message, throwable = throwable)
     }
 
@@ -71,22 +71,22 @@ class Logger(
         /**
          * Send a DEBUG log message.
          */
-        fun debug(message: String? = null, throwable: Throwable? = null) = DEFAULT.debug(message, throwable)
+        fun debug(message: String, throwable: Throwable? = null) = DEFAULT.debug(message, throwable)
 
         /**
          * Send a INFO log message.
          */
-        fun info(message: String? = null, throwable: Throwable? = null) = DEFAULT.info(message, throwable)
+        fun info(message: String, throwable: Throwable? = null) = DEFAULT.info(message, throwable)
 
         /**
          * Send a WARN log message.
          */
-        fun warn(message: String? = null, throwable: Throwable? = null) = DEFAULT.warn(message, throwable)
+        fun warn(message: String, throwable: Throwable? = null) = DEFAULT.warn(message, throwable)
 
         /**
          * Send a ERROR log message.
          */
-        fun error(message: String? = null, throwable: Throwable? = null) = DEFAULT.error(message, throwable)
+        fun error(message: String, throwable: Throwable? = null) = DEFAULT.error(message, throwable)
 
         /**
          * Measure the time it takes to execute the provided block and print a log message before and